Small tweaks

+++ kernel/src/lib/video.h
@@ -18,4 +18,4 @@ void video_clear(uint32_t color);
 void draw_rect(uint64_t x, uint64_t y, uint64_t w, uint64_t h, uint32_t color);
 void draw_filled_rect(uint64_t x, uint64_t y, uint64_t w, uint64_t h, uint32_t color);
 
-uint64_t color_from_argb(uint8_t a, uint8_t r, uint8_t g, uint8_t b);
+uint32_t color_from_argb(uint8_t a, uint8_t r, uint8_t g, uint8_t b);
